+++ include/flatcc/flatcc_version.h
@@ -2,12 +2,12 @@
 extern "C" {
 #endif
 
-#define FLATCC_VERSION_TEXT "0.5.2-pre"
+#define FLATCC_VERSION_TEXT "0.5.2"
 #define FLATCC_VERSION_MAJOR 0
 #define FLATCC_VERSION_MINOR 5
 #define FLATCC_VERSION_PATCH 2
 /* 1 or 0 */
-#define FLATCC_VERSION_RELEASED 0
+#define FLATCC_VERSION_RELEASED 1
 
 #ifdef __cplusplus
 }